+++ packages/eslint-plugin/src/utils/utils.ts
@@ -286,10 +286,10 @@ export function getNearestNodeFrom<T extends TSESTree.Node>(
 ): T | null {
   while (parent && !isProgram(parent)) {
     if (predicate(parent)) {
-      return (parent as unknown) as T;
+      return parent;
     }
 
-    parent = parent.parent as TSESTree.Node | undefined;
+    parent = parent.parent;
   }
 
   return null;
@@ -302,7 +302,7 @@ export function getImportDeclarations(
   let parentNode: TSESTree.Node | undefined = node;
 
   while ((parentNode = parentNode.parent)) {
-    if (parentNode.type !== 'Program') continue;
+    if (!isProgram(parentNode)) continue;
 
     return parentNode.body.filter(
       (node): node is TSESTree.ImportDeclaration =>
@@ -318,17 +318,17 @@ export function getImplementsRemoveFix(
   classDeclaration: TSESTree.ClassDeclaration,
   interfaceName: string,
   fixer: TSESLint.RuleFixer,
-): TSESLint.RuleFix | TSESLint.RuleFix[] {
+): TSESLint.RuleFix | undefined {
   const { implements: classImplements } = classDeclaration;
 
-  if (!classImplements) return [];
+  if (!classImplements) return undefined;
 
   const identifier = classImplements
     .map(({ expression }) => expression)
     .filter(isIdentifier)
     .find(({ name }) => name === interfaceName);
 
-  if (!identifier) return [];
+  if (!identifier) return undefined;
 
   const isFirstInterface = classImplements[0].expression === identifier;
   const isLastInterface =
@@ -353,14 +353,12 @@ export function getImplementsRemoveFix(
 
   const tokenBeforeInterface = sourceCode.getTokenBefore(identifier);
 
-  if (tokenBeforeInterface) {
-    return fixer.removeRange([
-      tokenBeforeInterface.range[0],
-      identifier.range[1],
-    ]);
-  }
+  if (!tokenBeforeInterface) return undefined;
 
-  return [];
+  return fixer.removeRange([
+    tokenBeforeInterface.range[0],
+    identifier.range[1],
+  ]);
 }
 
 export function getNodeToCommaRemoveFix(
@@ -397,16 +395,41 @@ function getImportDeclarationSpecifier(
   return undefined;
 }
 
+export function getLast<T extends readonly unknown[]>(items: T): T[number] {
+  return items.slice(-1)[0];
+}
+
+export function getImportAddFix(
+  node: TSESTree.ClassDeclaration,
+  moduleName: string,
+  importedName: string,
+  fixer: TSESLint.RuleFixer,
+): TSESLint.RuleFix {
+  const importDeclarations = getImportDeclarations(node, moduleName);
+
+  if (!importDeclarations?.length) {
+    return fixer.insertTextAfterRange(
+      [0, 0],
+      `import { ${importedName} } from '${moduleName}';\n`,
+    );
+  }
+
+  const firstImportDeclaration = importDeclarations[0];
+  const lastImportSpecifier = getLast(firstImportDeclaration.specifiers);
+
+  return fixer.insertTextAfter(lastImportSpecifier, `, ${importedName}`);
+}
+
 export function getImportRemoveFix(
   sourceCode: Readonly<TSESLint.SourceCode>,
   importDeclarations: readonly TSESTree.ImportDeclaration[],
   importedName: string,
   fixer: TSESLint.RuleFixer,
-): TSESLint.RuleFix | TSESLint.RuleFix[] {
+): TSESLint.RuleFix | undefined {
   const { importDeclaration, importSpecifier } =
     getImportDeclarationSpecifier(importDeclarations, importedName) ?? {};
 
-  if (!importDeclaration || !importSpecifier) return [];
+  if (!importDeclaration || !importSpecifier) return undefined;
 
   const isFirstImportSpecifier =
     importDeclaration.specifiers[0] === importSpecifier;
@@ -431,14 +454,28 @@ export function getImportRemoveFix(
 
   const tokenBeforeImportSpecifier = sourceCode.getTokenBefore(importSpecifier);
 
-  if (tokenBeforeImportSpecifier) {
-    return fixer.removeRange([
-      tokenBeforeImportSpecifier.range[0],
-      importSpecifier.range[1],
-    ]);
-  }
+  if (!tokenBeforeImportSpecifier) return undefined;
 
-  return [];
+  return fixer.removeRange([
+    tokenBeforeImportSpecifier.range[0],
+    importSpecifier.range[1],
+  ]);
+}
+
+export function getImplementsSchemaFixer(
+  { id, implements: implementz }: TSESTree.ClassDeclaration,
+  interfaceName: string,
+): {
+  readonly implementsNodeReplace:
+    | TSESTree.TSClassImplements
+    | TSESTree.Identifier;
+  readonly implementsTextReplace: string;
+} {
+  const [implementsNodeReplace, implementsTextReplace] = implementz
+    ? [getLast(implementz), `, ${interfaceName}`]
+    : [id as TSESTree.Identifier, ` implements ${interfaceName}`];
+
+  return { implementsNodeReplace, implementsTextReplace } as const;
 }
 
 export const getClassName = (node: TSESTree.Node): string | undefined => {
@@ -723,24 +760,20 @@ export const kebabToCamelCase = (value: string) =>
 
 export function isImportedFrom(
   identifier: TSESTree.Identifier,
-  module: string,
-) {
-  let parentNode: TSESTree.Node | undefined = identifier;
-  while ((parentNode = parentNode.parent)) {
-    if (parentNode.type !== 'Program') continue;
-    return parentNode.body.some(
-      (node) =>
-        isImportDeclaration(node) &&
-        (node.source as TSESTree.Literal).value === module &&
-        node.specifiers.some(
-          (specifier) =>
-            isImportSpecifier(specifier) &&
-            specifier.imported.name === identifier.name &&
-            specifier.local.name === identifier.name,
-        ),
-    );
-  }
-  return false;
+  moduleName: string,
+): boolean {
+  const importDeclarations = getImportDeclarations(identifier, moduleName);
+
+  return Boolean(
+    importDeclarations?.some((importDeclaration) =>
+      importDeclaration.specifiers.some(
+        (specifier) =>
+          isImportSpecifier(specifier) &&
+          specifier.imported.name === identifier.name &&
+          specifier.local.name === identifier.name,
+      ),
+    ),
+  );
 }
 
 /**